+++ packages/ccui/ui/date-picker/src/date-picker-types.ts
@@ -25,7 +25,7 @@ export const datePickerProps = {
   },
   placeholder: {
     type: String,
-    default: '请选择日期',
+    default: '',
   },
   disabled: {
     type: Boolean,